- DEVSU-2544 - Combine search views into one for ease of usage - Add signature variant data type and update dependant variant types - Fix coalesce and kb matches columns bug where sigv missing gene name property - Code lint
