+++ app/utils/searchReportsColumns.js
@@ -2,7 +2,7 @@ import startCase from '@/utils/startCase';
 
 function searchReportsColumns(report, analyst, reviewer, bioinformatician, searchType) {
   return {
-    matchedVariant: searchType=== 'kbmatches' ? report?.kbMatches[0]?.kbVariant : report?.genomicAlterationsIdentified[0]?.geneVariant,
+    matchedVariant: searchType === 'kbmatches' ? report?.kbMatches[0]?.kbVariant : report?.genomicAlterationsIdentified[0]?.geneVariant,
     patientID: report.patientId,
     analysisBiopsy: report.biopsyName,
     reportType: report.template.name === 'probe' ? 'Targeted Gene' : startCase(report.template.name),
